Created the two subclasses EnglishColumns and SpanishColumns, to handle the different functionality of the different decks
